@@ -3,28 +3,26 @@
   "version": "1.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --proxy-config proxy.conf.json",
-    "dev-integration": "ng serve --configuration integration --proxy-config proxy.conf.json",
+    "start": "ng serve",
+    "dev-integration": "ng serve --configuration integration",
     "build": "ng build --configuration production",
-    "integration": "ng build --configuration production --configuration integration",
+    "integration": "ng build --configuration integration",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.1.3",
-    "@angular/cdk": "^14.1.3",
-    "@angular/common": "^14.1.3",
-    "@angular/compiler": "^14.1.3",
-    "@angular/core": "^14.1.3",
-    "@angular/forms": "^14.1.3",
-    "@angular/material": "^14.1.3",
-    "@angular/platform-browser": "^14.1.3",
-    "@angular/platform-browser-dynamic": "^14.1.3",
-    "@angular/router": "^14.1.3",
-    "bootstrap": "^5.1.3",
-    "bootstrap-icons": "^1.6.1",
+    "@angular/animations": "^15.0.4",
+    "@angular/cdk": "^15.0.4",
+    "@angular/common": "^15.0.4",
+    "@angular/compiler": "^15.0.4",
+    "@angular/core": "^15.0.4",
+    "@angular/forms": "^15.0.4",
+    "@angular/material": "^15.0.4",
+    "@angular/platform-browser": "^15.0.4",
+    "@angular/platform-browser-dynamic": "^15.0.4",
+    "@angular/router": "^15.0.4",
     "font-awesome": "^4.7.0",
     "moment": "^2.29.4",
     "ng-lazyload-image": "^9.1.0",
@@ -33,10 +31,10 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.1.3",
-    "@angular/cli": "^14.1.3",
-    "@angular/compiler-cli": "^14.1.3",
-    "@angular/language-service": "^14.1.3",
+    "@angular-devkit/build-angular": "^15.0.5",
+    "@angular/cli": "^15.0.5",
+    "@angular/compiler-cli": "^15.0.4",
+    "@angular/language-service": "^15.0.4",
     "@types/jasmine": "~3.10.0",
     "@types/node": "~12.11.1",
     "codelyzer": "^6.0.0",
@@ -49,6 +47,6 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "ts-node": "~10.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.7.4"
+    "typescript": "~4.8.4"
   }
 }
